refactor(Delimiter) : 구분자 관련 클래스 재구성 및 패키지 분리
