+++ src/main/java/calculator/domain/delimiter/CustomDelimiter.java
@@ -1,4 +1,4 @@
-package calculator.domain;
+package calculator.domain.delimiter;
 
 public class CustomDelimiter {
